Refactor Rrd_reader and Rrd_writer
Instead of keeping a file descriptor as the state of a FileReader and
mmap'ing it each tick, mmap once and keep the resulting Cstruct as the
state.

This requires an interface change - the FileReader needs to know the
size of the file upfront, instead of being able to determine it during
the mmap of the file descriptor.

Signed-off-by: John Else <john.else@citrix.com>